Rename enums/ to enum/
